% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkmeans.R
\name{wkmeans}
\alias{wkmeans}
\title{Test}
\usage{
wkmeans(x, k, w = rep(1, nrow(x)), iter_max = 10, nstart = 1, cores = 1)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a matrix (such as a numeric vector or a data frame with all numeric columns).}

\item{k}{number of clusters.}

\item{w}{weights for each object (row) in x. By default, all observations are weighted the same, which yields regular kmeans.}

\item{iter_max}{maximum number of iterations allowed.}

\item{nstart}{number of random initialisation of clusters. Should be high for the clustering results to be stable.}

\item{cores}{numbers of cores to use in to try several initialisations in parallel (only useful if nstart > 1).}
}
\description{
Test
}
\examples{
# start with fake data
x <- matrix(runif(1000*2), ncol=2)
# compute and plot kmeans clusters
g <- wkmeans(x, k=4, iter_max=100, nstart=100)
plot(x, col=factor(g$cluster))
# put more weight on the left side (x < 0.5) and re-cluster
w <- rep(1, times=nrow(x))
w[x[,1]<0.5] <- 10
g <- wkmeans(x, k=4, iter_max=100, nstart=100, w=w)
plot(x, col=factor(g$cluster))
# more clusters are created on the left, as expected
}
